usize maths

+++ utils/zerovec/src/hashmap/mod.rs
@@ -68,11 +68,11 @@ where
         A: Borrow<K> + ?Sized,
     {
         let hash = compute_hash(key.borrow());
-        let (g, f0, f1) = split_hash64(hash, self.len() as u32);
+        let (g, f0, f1) = split_hash64(hash, self.len());
 
         #[allow(clippy::unwrap_used)] // g is in-range
         let (d0, d1) = self.displacements.get(g).unwrap();
-        let index = compute_index((f0, f1), (d0, d1), self.displacements.len() as u32)?;
+        let index = compute_index((f0, f1), (d0, d1), self.displacements.len())?;
 
         #[allow(clippy::unwrap_used)] // index is in 0..self.keys.len()
         let found = self.keys.zvl_get(index).unwrap();
